@@ -1,6 +1,7 @@
 package com.todo.dao;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+
 import java.time.LocalDateTime;
 
 @Entity
